+++ src/ibldsp/icsd.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-'''
+"""
 py-iCSD toolbox!
 Translation of the core functionality of the CSDplotter MATLAB package
 to python.
@@ -273,7 +273,7 @@ plt.show()
 
 
 
-'''
+"""
 
 import numpy as np
 import scipy.integrate as si
@@ -282,27 +282,31 @@ import quantities as pq
 import neo
 
 
-#patch quantities with the SI unit Siemens if it does not exist
+# patch quantities with the SI unit Siemens if it does not exist
 for symbol, prefix, definition, u_symbol in zip(
-    ['siemens', 'S', 'mS', 'uS', 'nS', 'pS'],
-    ['', '', 'milli', 'micro', 'nano', 'pico'],
-    [pq.A/pq.V, pq.A/pq.V, 'S', 'mS', 'uS', 'nS'],
-    [None, None, None, None, u'µS', None]):
+    ["siemens", "S", "mS", "uS", "nS", "pS"],
+    ["", "", "milli", "micro", "nano", "pico"],
+    [pq.A / pq.V, pq.A / pq.V, "S", "mS", "uS", "nS"],
+    [None, None, None, None, "µS", None],
+):
     if type(definition) is str:
         definition = lastdefinition / 1000
     if not hasattr(pq, symbol):
-        setattr(pq, symbol, pq.UnitQuantity(
-            prefix + 'siemens',
-            definition,
-            symbol=symbol,
-            u_symbol=u_symbol))
+        setattr(
+            pq,
+            symbol,
+            pq.UnitQuantity(
+                prefix + "siemens", definition, symbol=symbol, u_symbol=u_symbol
+            ),
+        )
     lastdefinition = definition
 
 
 class CSD(object):
-    '''Base iCSD class'''
-    def __init__(self, lfp, f_type='gaussian', f_order=(3, 1)):
-        '''Initialize parent class iCSD
+    """Base iCSD class"""
+
+    def __init__(self, lfp, f_type="gaussian", f_order=(3, 1)):
+        """Initialize parent class iCSD
 
         Parameters
         ----------
@@ -312,16 +316,17 @@ class CSD(object):
             type of spatial filter, must be a scipy.signal filter design method
         f_order : list
             settings for spatial filter, arg passed to  filter design function
-        '''
-        self.name = 'CSD estimate parent class'
+        """
+        self.name = "CSD estimate parent class"
         self.lfp = lfp
         self.f_matrix = np.eye(lfp.shape[0]) * pq.m**3 / pq.S
         self.f_type = f_type
         self.f_order = f_order
 
-
-    def get_csd(self, ):
-        '''
+    def get_csd(
+        self,
+    ):
+        """
         Perform the CSD estimate from the LFP and forward matrix F, i.e as
         CSD=F**-1*LFP
 
@@ -332,14 +337,13 @@ class CSD(object):
         -------
         csd : np.ndarray * quantity.Quantity
             Array with the csd estimate
-        '''
+        """
         csd = np.linalg.solve(self.f_matrix, self.lfp)
 
-        return csd * (self.f_matrix.units**-1*self.lfp.units).simplified
+        return csd * (self.f_matrix.units**-1 * self.lfp.units).simplified
 
-
-    def filter_csd(self, csd, filterfunction='convolve'):
-        '''
+    def filter_csd(self, csd, filterfunction="convolve"):
+        """
         Spatial filtering of the CSD estimate, using an N-point filter
 
         Arguments
@@ -349,71 +353,82 @@ class CSD(object):
         filterfunction : str
             'filtfilt' or 'convolve'. Apply spatial filter using
             scipy.signal.filtfilt or scipy.signal.convolve.
-        '''
-        if self.f_type == 'gaussian':
+        """
+        if self.f_type == "gaussian":
             try:
-                assert(len(self.f_order) == 2)
+                assert len(self.f_order) == 2
             except AssertionError as ae:
-                raise ae('filter order f_order must be a tuple of length 2')
+                raise ae("filter order f_order must be a tuple of length 2")
         else:
             try:
-                assert(self.f_order > 0 and isinstance(self.f_order, int))
+                assert self.f_order > 0 and isinstance(self.f_order, int)
             except AssertionError as ae:
-                raise ae('Filter order must be int > 0!')
+                raise ae("Filter order must be int > 0!")
         try:
-            assert(filterfunction in ['filtfilt', 'convolve'])
+            assert filterfunction in ["filtfilt", "convolve"]
         except AssertionError as ae:
             raise ae("{} not equal to 'filtfilt' or 'convolve'".format(filterfunction))
 
-        if self.f_type == 'boxcar':
+        if self.f_type == "boxcar":
             num = ss.boxcar(self.f_order)
             denom = np.array([num.sum()])
-        elif self.f_type == 'hamming':
+        elif self.f_type == "hamming":
             num = ss.hamming(self.f_order)
             denom = np.array([num.sum()])
-        elif self.f_type == 'triangular':
+        elif self.f_type == "triangular":
             num = ss.triang(self.f_order)
             denom = np.array([num.sum()])
-        elif self.f_type == 'gaussian':
+        elif self.f_type == "gaussian":
             num = ss.gaussian(self.f_order[0], self.f_order[1])
             denom = np.array([num.sum()])
-        elif self.f_type == 'identity':
-            num = np.array([1.])
-            denom = np.array([1.])
+        elif self.f_type == "identity":
+            num = np.array([1.0])
+            denom = np.array([1.0])
         else:
-            print('%s Wrong filter type!' % self.f_type)
+            print("%s Wrong filter type!" % self.f_type)
             raise
 
-        num_string = '[ '
+        num_string = "[ "
         for i in num:
-            num_string = num_string + '%.3f ' % i
-        num_string = num_string + ']'
-        denom_string = '[ '
+            num_string = num_string + "%.3f " % i
+        num_string = num_string + "]"
+        denom_string = "[ "
         for i in denom:
-            denom_string = denom_string + '%.3f ' % i
-        denom_string = denom_string + ']'
+            denom_string = denom_string + "%.3f " % i
+        denom_string = denom_string + "]"
 
-        print(('discrete filter coefficients: \nb = {}, \na = {}'.format(num_string, denom_string)))
+        print(
+            (
+                "discrete filter coefficients: \nb = {}, \na = {}".format(
+                    num_string, denom_string
+                )
+            )
+        )
 
-        if filterfunction == 'filtfilt':
+        if filterfunction == "filtfilt":
             return ss.filtfilt(num, denom, csd, axis=0) * csd.units
-        elif filterfunction == 'convolve':
+        elif filterfunction == "convolve":
             csdf = csd / csd.units
             for i in range(csdf.shape[1]):
-                csdf[:, i] = ss.convolve(csdf[:, i], num/denom.sum(), 'same')
+                csdf[:, i] = ss.convolve(csdf[:, i], num / denom.sum(), "same")
             return csdf * csd.units
 
 
 class StandardCSD(CSD):
-    '''
+    """
     Standard CSD method with and without Vaknin electrodes
-    '''
+    """
 
-    def __init__(self, lfp,
-                 coord_electrode=np.linspace(0, 1400E-6, 15)*pq.m,
-                 sigma=0.3*pq.S/pq.m, vaknin_el=True, f_type='gaussian',
-                 f_order=(3, 1)):
-        '''
+    def __init__(
+        self,
+        lfp,
+        coord_electrode=np.linspace(0, 1400e-6, 15) * pq.m,
+        sigma=0.3 * pq.S / pq.m,
+        vaknin_el=True,
+        f_type="gaussian",
+        f_order=(3, 1),
+    ):
+        """
         Initialize standard CSD method class with and without Vaknin electrodes.
 
         Parameters
@@ -431,58 +446,56 @@ class StandardCSD(CSD):
             type of spatial filter, must be a scipy.signal filter design method
         f_order : list
             settings for spatial filter, arg passed to  filter design function
-        '''
+        """
         CSD.__init__(self, lfp, f_type, f_order)
 
-        self.name = 'Standard CSD method'
+        self.name = "Standard CSD method"
         self.coord_electrode = coord_electrode
         self.sigma = sigma
         self.vaknin_el = vaknin_el
 
         try:
-            assert(np.all(np.diff(np.diff(coord_electrode)))==0)
+            assert np.all(np.diff(np.diff(coord_electrode))) == 0
         except AssertionError as ae:
-            print('coord_electrode not monotonously varying')
+            print("coord_electrode not monotonously varying")
             raise ae
 
         if vaknin_el:
-            #extend array of lfps by duplicating potential at endpoint contacts
+            # extend array of lfps by duplicating potential at endpoint contacts
             if lfp.ndim == 1:
-                self.lfp = np.empty((lfp.shape[0]+2, )) * lfp.units
+                self.lfp = np.empty((lfp.shape[0] + 2,)) * lfp.units
             else:
-                self.lfp = np.empty((lfp.shape[0]+2, lfp.shape[1])) * lfp.units
-            self.lfp[0, ] = lfp[0, ]
-            self.lfp[1:-1, ] = lfp
-            self.lfp[-1, ] = lfp[-1, ]
+                self.lfp = np.empty((lfp.shape[0] + 2, lfp.shape[1])) * lfp.units
+            self.lfp[0,] = lfp[0,]
+            self.lfp[1:-1,] = lfp
+            self.lfp[-1,] = lfp[-1,]
         else:
             self.lfp = lfp
 
         self.f_inv_matrix = self.get_f_inv_matrix()
 
-
     def get_f_inv_matrix(self):
-        '''Calculate the inverse F-matrix for the standard CSD method'''
+        """Calculate the inverse F-matrix for the standard CSD method"""
         h_val = abs(np.diff(self.coord_electrode)[0])
 
         f_inv = -np.eye(self.lfp.shape[0])
 
-        #Inner matrix elements  is just the discrete laplacian coefficients
-        for j in range(1, f_inv.shape[0]-1):
-            f_inv[j, j-1:j+2] = np.array([1., -2., 1.])
+        # Inner matrix elements  is just the discrete laplacian coefficients
+        for j in range(1, f_inv.shape[0] - 1):
+            f_inv[j, j - 1 : j + 2] = np.array([1.0, -2.0, 1.0])
 
         return f_inv * -self.sigma / h_val
 
-
     def get_csd(self):
-        '''
+        """
         Perform the iCSD calculation, i.e: iCSD=F_inv*LFP
 
         Returns
         -------
         csd : np.ndarray * quantity.Quantity
             Array with the csd estimate
-        '''
-        csd = np.dot(self.f_inv_matrix, self.lfp)[1:-1, ]
+        """
+        csd = np.dot(self.f_inv_matrix, self.lfp)[1:-1,]
         # `np.dot()` does not return correct units, so the units of `csd` must
         # be assigned manually
         csd_units = (self.f_inv_matrix.units * self.lfp.units).simplified
@@ -492,16 +505,21 @@ class StandardCSD(CSD):
 
 
 class DeltaiCSD(CSD):
-    '''
+    """
     delta-iCSD method
-    '''
-    def __init__(self, lfp,
-                 coord_electrode=np.linspace(0, 1400E-6, 15)*pq.m,
-                 diam=500E-6*pq.m,
-                 sigma=0.3*pq.S/pq.m,
-                 sigma_top=0.3*pq.S/pq.m,
-                 f_type='gaussian', f_order=(3, 1)):
-        '''
+    """
+
+    def __init__(
+        self,
+        lfp,
+        coord_electrode=np.linspace(0, 1400e-6, 15) * pq.m,
+        diam=500e-6 * pq.m,
+        sigma=0.3 * pq.S / pq.m,
+        sigma_top=0.3 * pq.S / pq.m,
+        f_type="gaussian",
+        f_order=(3, 1),
+    ):
+        """
         Initialize the delta-iCSD method class object
 
         Parameters
@@ -523,35 +541,38 @@ class DeltaiCSD(CSD):
         f_order : list
             settings for spatial filter, arg passed to  filter design function
 
-        '''
+        """
         CSD.__init__(self, lfp)
 
         try:
-            assert(diam.units == coord_electrode.units)
+            assert diam.units == coord_electrode.units
         except AssertionError as ae:
-            print('units of coord_electrode ({}) and diam ({}) differ'.format(coord_electrode.units,
-                                                                                  diam.units))
+            print(
+                "units of coord_electrode ({}) and diam ({}) differ".format(
+                    coord_electrode.units, diam.units
+                )
+            )
             raise ae
 
         try:
-            assert(np.all(np.diff(coord_electrode) > 0))
+            assert np.all(np.diff(coord_electrode) > 0)
         except AssertionError as ae:
-            print('values of coord_electrode not continously increasing')
+            print("values of coord_electrode not continously increasing")
             raise ae
 
         try:
-            assert(diam.size == 1 or diam.size == coord_electrode.size)
+            assert diam.size == 1 or diam.size == coord_electrode.size
             if diam.size == coord_electrode.size:
-                assert(np.all(diam > 0*diam.units))
+                assert np.all(diam > 0 * diam.units)
             else:
-                assert(diam > 0*diam.units)
+                assert diam > 0 * diam.units
         except AssertionError as ae:
-            print('diam must be positive scalar or of same shape as coord_electrode')
+            print("diam must be positive scalar or of same shape as coord_electrode")
             raise ae
         if diam.size == 1:
-            diam = np.ones(coord_electrode.size)*diam
+            diam = np.ones(coord_electrode.size) * diam
 
-        self.name = 'delta-iCSD method'
+        self.name = "delta-iCSD method"
         self.coord_electrode = coord_electrode
         self.diam = diam
         self.sigma = sigma
@@ -559,43 +580,53 @@ class DeltaiCSD(CSD):
         self.f_type = f_type
         self.f_order = f_order
 
-        #initialize F- and iCSD-matrices
-        self.f_matrix = np.empty((self.coord_electrode.size,
-                                  self.coord_electrode.size))
+        # initialize F- and iCSD-matrices
+        self.f_matrix = np.empty((self.coord_electrode.size, self.coord_electrode.size))
 
         self.f_matrix = self.get_f_matrix()
 
-
     def get_f_matrix(self):
-        '''Calculate the F-matrix'''
-        f_matrix = np.empty((self.coord_electrode.size,
-                             self.coord_electrode.size))*self.coord_electrode.units
+        """Calculate the F-matrix"""
+        f_matrix = (
+            np.empty((self.coord_electrode.size, self.coord_electrode.size))
+            * self.coord_electrode.units
+        )
         for j in range(self.coord_electrode.size):
             for i in range(self.coord_electrode.size):
-                f_matrix[j, i] = ((np.sqrt((self.coord_electrode[j] -
-                                            self.coord_electrode[i])**2 +
-                    (self.diam[j] / 2)**2) - abs(self.coord_electrode[j] -
-                                              self.coord_electrode[i])) +
-                    (self.sigma - self.sigma_top) / (self.sigma +
-                                                     self.sigma_top) *
-                    (np.sqrt((self.coord_electrode[j] +
-                              self.coord_electrode[i])**2 + (self.diam[j] / 2)**2)-
-                    abs(self.coord_electrode[j] + self.coord_electrode[i])))
+                f_matrix[j, i] = (
+                    np.sqrt(
+                        (self.coord_electrode[j] - self.coord_electrode[i]) ** 2
+                        + (self.diam[j] / 2) ** 2
+                    )
+                    - abs(self.coord_electrode[j] - self.coord_electrode[i])
+                ) + (self.sigma - self.sigma_top) / (self.sigma + self.sigma_top) * (
+                    np.sqrt(
+                        (self.coord_electrode[j] + self.coord_electrode[i]) ** 2
+                        + (self.diam[j] / 2) ** 2
+                    )
+                    - abs(self.coord_electrode[j] + self.coord_electrode[i])
+                )
 
-        f_matrix /= (2 * self.sigma)
+        f_matrix /= 2 * self.sigma
         return f_matrix
 
 
 class StepiCSD(CSD):
-    '''step-iCSD method'''
-    def __init__(self, lfp,
-                 coord_electrode=np.linspace(0, 1400E-6, 15)*pq.m,
-                 diam=500E-6*pq.m,
-                 h = np.ones(15)*100E-6*pq.m,
-                 sigma=0.3*pq.S/pq.m, sigma_top=0.3*pq.S/pq.m,
-                 tol=1E-6,
-                 f_type='gaussian', f_order=(3, 1)):
-        '''
+    """step-iCSD method"""
+
+    def __init__(
+        self,
+        lfp,
+        coord_electrode=np.linspace(0, 1400e-6, 15) * pq.m,
+        diam=500e-6 * pq.m,
+        h=np.ones(15) * 100e-6 * pq.m,
+        sigma=0.3 * pq.S / pq.m,
+        sigma_top=0.3 * pq.S / pq.m,
+        tol=1e-6,
+        f_type="gaussian",
+        f_order=(3, 1),
+    ):
+        """
         Initializing step-iCSD method class object
 
         Parameters
@@ -621,45 +652,47 @@ class StepiCSD(CSD):
         f_order : list
             settings for spatial filter, arg passed to  filter design function
 
-        '''
+        """
         CSD.__init__(self, lfp, f_type, f_order)
 
         try:
-            assert(diam.units == coord_electrode.units)
+            assert diam.units == coord_electrode.units
         except AssertionError as ae:
-            print('units of coord_electrode ({}) and diam ({}) differ'.format(coord_electrode.units,
-                                                                                  diam.units))
+            print(
+                "units of coord_electrode ({}) and diam ({}) differ".format(
+                    coord_electrode.units, diam.units
+                )
+            )
             raise ae
         try:
-            assert(np.all(np.diff(coord_electrode) > 0))
+            assert np.all(np.diff(coord_electrode) > 0)
         except AssertionError as ae:
-            print('values of coord_electrode not continously increasing')
+            print("values of coord_electrode not continously increasing")
             raise ae
 
         try:
-            assert(diam.size == 1 or diam.size == coord_electrode.size)
+            assert diam.size == 1 or diam.size == coord_electrode.size
             if diam.size == coord_electrode.size:
-                assert(np.all(diam > 0*diam.units))
+                assert np.all(diam > 0 * diam.units)
             else:
-                assert(diam > 0*diam.units)
+                assert diam > 0 * diam.units
         except AssertionError as ae:
-            print('diam must be positive scalar or of same shape as coord_electrode')
+            print("diam must be positive scalar or of same shape as coord_electrode")
             raise ae
         if diam.size == 1:
-            diam = np.ones(coord_electrode.size)*diam
+            diam = np.ones(coord_electrode.size) * diam
 
         try:
-            assert(h.size == 1 or h.size == coord_electrode.size)
+            assert h.size == 1 or h.size == coord_electrode.size
             if h.size == coord_electrode.size:
-                assert(np.all(h > 0*h.units))
+                assert np.all(h > 0 * h.units)
         except AssertionError as ae:
-            print('h must be scalar or of same shape as coord_electrode')
+            print("h must be scalar or of same shape as coord_electrode")
             raise ae
         if h.size == 1:
-            h = np.ones(coord_electrode.size)*h
+            h = np.ones(coord_electrode.size) * h
 
-
-        self.name = 'step-iCSD method'
+        self.name = "step-iCSD method"
         self.coord_electrode = coord_electrode
         self.diam = diam
         self.h = h
@@ -670,54 +703,75 @@ class StepiCSD(CSD):
         # compute forward-solution matrix
         self.f_matrix = self.get_f_matrix()
 
-
     def get_f_matrix(self):
-        '''Calculate F-matrix for step iCSD method'''
+        """Calculate F-matrix for step iCSD method"""
         el_len = self.coord_electrode.size
         f_matrix = np.zeros((el_len, el_len))
         for j in range(el_len):
             for i in range(el_len):
-                lower_int = self.coord_electrode[i] - self.h[j]/2
+                lower_int = self.coord_electrode[i] - self.h[j] / 2
                 if lower_int < 0:
                     lower_int = self.h[j].units
                 upper_int = self.coord_electrode[i] + self.h[j] / 2
 
-                #components of f_matrix object
-                f_cyl0 = si.quad(self._f_cylinder,
-                                 a=lower_int, b=upper_int,
-                                 args=(float(self.coord_electrode[j]),
-                                       float(self.diam[j]),
-                                       float(self.sigma)),
-                                 epsabs=self.tol)[0]
-                f_cyl1 = si.quad(self._f_cylinder, a=lower_int, b=upper_int,
-                                 args=(-float(self.coord_electrode[j]),
-                                       float(self.diam[j]), float(self.sigma)),
-                                 epsabs=self.tol)[0]
+                # components of f_matrix object
+                f_cyl0 = si.quad(
+                    self._f_cylinder,
+                    a=lower_int,
+                    b=upper_int,
+                    args=(
+                        float(self.coord_electrode[j]),
+                        float(self.diam[j]),
+                        float(self.sigma),
+                    ),
+                    epsabs=self.tol,
+                )[0]
+                f_cyl1 = si.quad(
+                    self._f_cylinder,
+                    a=lower_int,
+                    b=upper_int,
+                    args=(
+                        -float(self.coord_electrode[j]),
+                        float(self.diam[j]),
+                        float(self.sigma),
+                    ),
+                    epsabs=self.tol,
+                )[0]
 
-                #method of images coefficient
-                mom = (self.sigma-self.sigma_top)/(self.sigma+self.sigma_top)
+                # method of images coefficient
+                mom = (self.sigma - self.sigma_top) / (self.sigma + self.sigma_top)
 
-                f_matrix[j, i] = f_cyl0 + mom*f_cyl1
+                f_matrix[j, i] = f_cyl0 + mom * f_cyl1
 
-        #assume si.quad trash the units
+        # assume si.quad trash the units
         return f_matrix * self.h.units**2 / self.sigma.units
 
-
     def _f_cylinder(self, zeta, z_val, diam, sigma):
-        '''function used by class method'''
-        f_cyl = 1. / (2.*sigma) * \
-            (np.sqrt((diam/2)**2 + ((z_val-zeta))**2) - abs(z_val-zeta))
+        """function used by class method"""
+        f_cyl = (
+            1.0
+            / (2.0 * sigma)
+            * (np.sqrt((diam / 2) ** 2 + (z_val - zeta) ** 2) - abs(z_val - zeta))
+        )
         return f_cyl
 
 
 class SplineiCSD(CSD):
-    '''spline iCSD method'''
-    def __init__(self, lfp,
-                 coord_electrode=np.linspace(0, 1400E-6, 15)*pq.m,
-                 diam=500E-6*pq.m, sigma=0.3*pq.S/pq.m, sigma_top=0.3*pq.S/pq.m,
-                 tol=1E-6,
-                 f_type='gaussian', f_order=(3, 1), num_steps=200):
-        '''
+    """spline iCSD method"""
+
+    def __init__(
+        self,
+        lfp,
+        coord_electrode=np.linspace(0, 1400e-6, 15) * pq.m,
+        diam=500e-6 * pq.m,
+        sigma=0.3 * pq.S / pq.m,
+        sigma_top=0.3 * pq.S / pq.m,
+        tol=1e-6,
+        f_type="gaussian",
+        f_order=(3, 1),
+        num_steps=200,
+    ):
+        """
         Initializing spline-iCSD method class object
 
         Parameters
@@ -743,32 +797,35 @@ class SplineiCSD(CSD):
         num_steps : int
             number of data points for the spatially upsampled LFP/CSD data
 
-        '''
+        """
         CSD.__init__(self, lfp, f_type, f_order)
 
         try:
-            assert(diam.units == coord_electrode.units)
+            assert diam.units == coord_electrode.units
         except AssertionError as ae:
-            print('units of coord_electrode ({}) and diam ({}) differ'.format(coord_electrode.units,
-                                                                                  diam.units))
+            print(
+                "units of coord_electrode ({}) and diam ({}) differ".format(
+                    coord_electrode.units, diam.units
+                )
+            )
             raise
         try:
-            assert(np.all(np.diff(coord_electrode) > 0))
+            assert np.all(np.diff(coord_electrode) > 0)
         except AssertionError as ae:
-            print('values of coord_electrode not continously increasing')
+            print("values of coord_electrode not continously increasing")
             raise ae
 
         try:
-            assert(diam.size == 1 or diam.size == coord_electrode.size)
+            assert diam.size == 1 or diam.size == coord_electrode.size
             if diam.size == coord_electrode.size:
-                assert(np.all(diam > 0*diam.units))
+                assert np.all(diam > 0 * diam.units)
         except AssertionError as ae:
-            print('diam must be scalar or of same shape as coord_electrode')
+            print("diam must be scalar or of same shape as coord_electrode")
             raise ae
         if diam.size == 1:
-            diam = np.ones(coord_electrode.size)*diam
+            diam = np.ones(coord_electrode.size) * diam
 
-        self.name = 'spline-iCSD method'
+        self.name = "spline-iCSD method"
         self.coord_electrode = coord_electrode
         self.diam = diam
         self.sigma = sigma
@@ -779,82 +836,132 @@ class SplineiCSD(CSD):
         # compute stuff
         self.f_matrix = self.get_f_matrix()
 
-
     def get_f_matrix(self):
-        '''Calculate the F-matrix for cubic spline iCSD method'''
+        """Calculate the F-matrix for cubic spline iCSD method"""
         el_len = self.coord_electrode.size
-        z_js = np.zeros(el_len+1)
+        z_js = np.zeros(el_len + 1)
         z_js[:-1] = np.array(self.coord_electrode)
         z_js[-1] = z_js[-2] + float(np.diff(self.coord_electrode).mean())
 
         # Define integration matrixes
-        f_mat0 = np.zeros((el_len, el_len+1))
-        f_mat1 = np.zeros((el_len, el_len+1))
-        f_mat2 = np.zeros((el_len, el_len+1))
-        f_mat3 = np.zeros((el_len, el_len+1))
+        f_mat0 = np.zeros((el_len, el_len + 1))
+        f_mat1 = np.zeros((el_len, el_len + 1))
+        f_mat2 = np.zeros((el_len, el_len + 1))
+        f_mat3 = np.zeros((el_len, el_len + 1))
 
         # Calc. elements
         for j in range(el_len):
             for i in range(el_len):
-                f_mat0[j, i] = si.quad(self._f_mat0, a=z_js[i], b=z_js[i+1],
-                                       args=(z_js[j+1],
-                                             float(self.sigma),
-                                             float(self.diam[j])),
-                                       epsabs=self.tol)[0]
-                f_mat1[j, i] = si.quad(self._f_mat1, a=z_js[i], b=z_js[i+1],
-                                       args=(z_js[j+1], z_js[i],
-                                             float(self.sigma),
-                                             float(self.diam[j])),
-                                       epsabs=self.tol)[0]
-                f_mat2[j, i] = si.quad(self._f_mat2, a=z_js[i], b=z_js[i+1],
-                                       args=(z_js[j+1], z_js[i],
-                                             float(self.sigma),
-                                             float(self.diam[j])),
-                                       epsabs=self.tol)[0]
-                f_mat3[j, i] = si.quad(self._f_mat3, a=z_js[i], b=z_js[i+1],
-                                       args=(z_js[j+1], z_js[i],
-                                             float(self.sigma),
-                                             float(self.diam[j])),
-                                       epsabs=self.tol)[0]
+                f_mat0[j, i] = si.quad(
+                    self._f_mat0,
+                    a=z_js[i],
+                    b=z_js[i + 1],
+                    args=(z_js[j + 1], float(self.sigma), float(self.diam[j])),
+                    epsabs=self.tol,
+                )[0]
+                f_mat1[j, i] = si.quad(
+                    self._f_mat1,
+                    a=z_js[i],
+                    b=z_js[i + 1],
+                    args=(z_js[j + 1], z_js[i], float(self.sigma), float(self.diam[j])),
+                    epsabs=self.tol,
+                )[0]
+                f_mat2[j, i] = si.quad(
+                    self._f_mat2,
+                    a=z_js[i],
+                    b=z_js[i + 1],
+                    args=(z_js[j + 1], z_js[i], float(self.sigma), float(self.diam[j])),
+                    epsabs=self.tol,
+                )[0]
+                f_mat3[j, i] = si.quad(
+                    self._f_mat3,
+                    a=z_js[i],
+                    b=z_js[i + 1],
+                    args=(z_js[j + 1], z_js[i], float(self.sigma), float(self.diam[j])),
+                    epsabs=self.tol,
+                )[0]
 
                 # image technique if conductivity not constant:
                 if self.sigma != self.sigma_top:
-                    f_mat0[j, i] = f_mat0[j, i] + (self.sigma-self.sigma_top) / \
-                                                (self.sigma + self.sigma_top) * \
-                            si.quad(self._f_mat0, a=z_js[i], b=z_js[i+1], \
-                                    args=(-z_js[j+1],
-                                          float(self.sigma), float(self.diam[j])), \
-                                    epsabs=self.tol)[0]
-                    f_mat1[j, i] = f_mat1[j, i] + (self.sigma-self.sigma_top) / \
-                        (self.sigma + self.sigma_top) * \
-                            si.quad(self._f_mat1, a=z_js[i], b=z_js[i+1], \
-                                args=(-z_js[j+1], z_js[i], float(self.sigma),
-                                      float(self.diam[j])), epsabs=self.tol)[0]
-                    f_mat2[j, i] = f_mat2[j, i] + (self.sigma-self.sigma_top) / \
-                        (self.sigma + self.sigma_top) * \
-                            si.quad(self._f_mat2, a=z_js[i], b=z_js[i+1], \
-                                args=(-z_js[j+1], z_js[i], float(self.sigma),
-                                      float(self.diam[j])), epsabs=self.tol)[0]
-                    f_mat3[j, i] = f_mat3[j, i] + (self.sigma-self.sigma_top) / \
-                        (self.sigma + self.sigma_top) * \
-                            si.quad(self._f_mat3, a=z_js[i], b=z_js[i+1], \
-                                args=(-z_js[j+1], z_js[i], float(self.sigma),
-                                      float(self.diam[j])), epsabs=self.tol)[0]
+                    f_mat0[j, i] = (
+                        f_mat0[j, i]
+                        + (self.sigma - self.sigma_top)
+                        / (self.sigma + self.sigma_top)
+                        * si.quad(
+                            self._f_mat0,
+                            a=z_js[i],
+                            b=z_js[i + 1],
+                            args=(-z_js[j + 1], float(self.sigma), float(self.diam[j])),
+                            epsabs=self.tol,
+                        )[0]
+                    )
+                    f_mat1[j, i] = (
+                        f_mat1[j, i]
+                        + (self.sigma - self.sigma_top)
+                        / (self.sigma + self.sigma_top)
+                        * si.quad(
+                            self._f_mat1,
+                            a=z_js[i],
+                            b=z_js[i + 1],
+                            args=(
+                                -z_js[j + 1],
+                                z_js[i],
+                                float(self.sigma),
+                                float(self.diam[j]),
+                            ),
+                            epsabs=self.tol,
+                        )[0]
+                    )
+                    f_mat2[j, i] = (
+                        f_mat2[j, i]
+                        + (self.sigma - self.sigma_top)
+                        / (self.sigma + self.sigma_top)
+                        * si.quad(
+                            self._f_mat2,
+                            a=z_js[i],
+                            b=z_js[i + 1],
+                            args=(
+                                -z_js[j + 1],
+                                z_js[i],
+                                float(self.sigma),
+                                float(self.diam[j]),
+                            ),
+                            epsabs=self.tol,
+                        )[0]
+                    )
+                    f_mat3[j, i] = (
+                        f_mat3[j, i]
+                        + (self.sigma - self.sigma_top)
+                        / (self.sigma + self.sigma_top)
+                        * si.quad(
+                            self._f_mat3,
+                            a=z_js[i],
+                            b=z_js[i + 1],
+                            args=(
+                                -z_js[j + 1],
+                                z_js[i],
+                                float(self.sigma),
+                                float(self.diam[j]),
+                            ),
+                            epsabs=self.tol,
+                        )[0]
+                    )
 
         e_mat0, e_mat1, e_mat2, e_mat3 = self._calc_e_matrices()
 
         # Calculate the F-matrix
-        f_matrix = np.eye(el_len+2)
-        f_matrix[1:-1, :] = np.dot(f_mat0, e_mat0) + \
-                            np.dot(f_mat1, e_mat1) + \
-                            np.dot(f_mat2, e_mat2) + \
-                            np.dot(f_mat3, e_mat3)
+        f_matrix = np.eye(el_len + 2)
+        f_matrix[1:-1, :] = (
+            np.dot(f_mat0, e_mat0)
+            + np.dot(f_mat1, e_mat1)
+            + np.dot(f_mat2, e_mat2)
+            + np.dot(f_mat3, e_mat3)
+        )
 
         return f_matrix * self.coord_electrode.units**2 / self.sigma.units
 
-
     def get_csd(self):
-        '''
+        """
         Calculate the iCSD using the spline iCSD method
 
         Returns
@@ -863,7 +970,7 @@ class SplineiCSD(CSD):
             Array with csd estimate
 
 
-        '''
+        """
         e_mat = self._calc_e_matrices()
 
         el_len = self.coord_electrode.size
@@ -873,9 +980,13 @@ class SplineiCSD(CSD):
             cs_lfp = np.r_[[0], np.asarray(self.lfp), [0]].reshape(1, -1).T
             csd = np.zeros(self.num_steps)
         else:
-            cs_lfp = np.vstack((np.zeros(self.lfp.shape[1]),
-                                np.asarray(self.lfp),
-                                np.zeros(self.lfp.shape[1])))
+            cs_lfp = np.vstack(
+                (
+                    np.zeros(self.lfp.shape[1]),
+                    np.asarray(self.lfp),
+                    np.zeros(self.lfp.shape[1]),
+                )
+            )
             csd = np.zeros((self.num_steps, self.lfp.shape[1]))
         cs_lfp *= self.lfp.units
 
@@ -901,99 +1012,120 @@ class SplineiCSD(CSD):
         # Calculate iCSD estimate on grid from polynomial coefficients.
         i = 0
         for j in range(self.num_steps):
-            if out_zs[j] >= z_js[i+1]:
+            if out_zs[j] >= z_js[i + 1]:
                 i += 1
-            csd[j,] = a_mat0[i, :] + a_mat1[i, :] * \
-                            (out_zs[j] - z_js[i]) +\
-                a_mat2[i, :] * (out_zs[j] - z_js[i])**2 + \
-                a_mat3[i, :] * (out_zs[j] - z_js[i])**3
+            csd[j,] = (
+                a_mat0[i, :]
+                + a_mat1[i, :] * (out_zs[j] - z_js[i])
+                + a_mat2[i, :] * (out_zs[j] - z_js[i]) ** 2
+                + a_mat3[i, :] * (out_zs[j] - z_js[i]) ** 3
+            )
 
         csd_unit = (self.f_matrix.units**-1 * self.lfp.units).simplified
 
         return csd * csd_unit
 
     def _f_mat0(self, zeta, z_val, sigma, diam):
-        '''0'th order potential function'''
-        return 1./(2.*sigma) * \
-            (np.sqrt((diam/2)**2 + ((z_val-zeta))**2) - abs(z_val-zeta))
+        """0'th order potential function"""
+        return (
+            1.0
+            / (2.0 * sigma)
+            * (np.sqrt((diam / 2) ** 2 + (z_val - zeta) ** 2) - abs(z_val - zeta))
+        )
 
     def _f_mat1(self, zeta, z_val, zi_val, sigma, diam):
-        '''1'th order potential function'''
-        return (zeta-zi_val) * self._f_mat0(zeta, z_val, sigma, diam)
+        """1'th order potential function"""
+        return (zeta - zi_val) * self._f_mat0(zeta, z_val, sigma, diam)
 
     def _f_mat2(self, zeta, z_val, zi_val, sigma, diam):
-        '''2'nd order potential function'''
-        return (zeta-zi_val)**2 * self._f_mat0(zeta, z_val, sigma, diam)
+        """2'nd order potential function"""
+        return (zeta - zi_val) ** 2 * self._f_mat0(zeta, z_val, sigma, diam)
 
     def _f_mat3(self, zeta, z_val, zi_val, sigma, diam):
-        '''3'rd order potential function'''
-        return (zeta-zi_val)**3 * self._f_mat0(zeta, z_val, sigma, diam)
-
+        """3'rd order potential function"""
+        return (zeta - zi_val) ** 3 * self._f_mat0(zeta, z_val, sigma, diam)
 
     def _calc_k_matrix(self):
-        '''Calculate the K-matrix used by to calculate E-matrices'''
+        """Calculate the K-matrix used by to calculate E-matrices"""
         el_len = self.coord_electrode.size
         h = float(np.diff(self.coord_electrode).min())
 
-        c_jm1 = np.eye(el_len+2, k=0) / h
+        c_jm1 = np.eye(el_len + 2, k=0) / h
         c_jm1[0, 0] = 0
 
-        c_j0 = np.eye(el_len+2) / h
+        c_j0 = np.eye(el_len + 2) / h
         c_j0[-1, -1] = 0
 
         c_jall = c_j0
         c_jall[0, 0] = 1
         c_jall[-1, -1] = 1
 
-        tjp1 = np.eye(el_len+2, k=1)
-        tjm1 = np.eye(el_len+2, k=-1)
+        tjp1 = np.eye(el_len + 2, k=1)
+        tjm1 = np.eye(el_len + 2, k=-1)
 
-        tj0 = np.eye(el_len+2)
+        tj0 = np.eye(el_len + 2)
         tj0[0, 0] = 0
         tj0[-1, -1] = 0
 
         # Defining K-matrix used to calculate e_mat1-3
-        return np.dot(np.linalg.inv(np.dot(c_jm1, tjm1) +
-                                    2*np.dot(c_jm1, tj0) +
-                                    2*c_jall +
-                                    np.dot(c_j0, tjp1)),
-                      3*(np.dot(np.dot(c_jm1, c_jm1), tj0) -
-                         np.dot(np.dot(c_jm1, c_jm1), tjm1) +
-                         np.dot(np.dot(c_j0, c_j0), tjp1) -
-                         np.dot(np.dot(c_j0, c_j0), tj0)))
-
+        return np.dot(
+            np.linalg.inv(
+                np.dot(c_jm1, tjm1)
+                + 2 * np.dot(c_jm1, tj0)
+                + 2 * c_jall
+                + np.dot(c_j0, tjp1)
+            ),
+            3
+            * (
+                np.dot(np.dot(c_jm1, c_jm1), tj0)
+                - np.dot(np.dot(c_jm1, c_jm1), tjm1)
+                + np.dot(np.dot(c_j0, c_j0), tjp1)
+                - np.dot(np.dot(c_j0, c_j0), tj0)
+            ),
+        )
 
     def _calc_e_matrices(self):
-        '''Calculate the E-matrices used by cubic spline iCSD method'''
+        """Calculate the E-matrices used by cubic spline iCSD method"""
         el_len = self.coord_electrode.size
         ## expanding electrode grid
         h = float(np.diff(self.coord_electrode).min())
 
         ## Define transformation matrices
-        c_mat3 = np.eye(el_len+1) / h
+        c_mat3 = np.eye(el_len + 1) / h
 
         # Get K-matrix
         k_matrix = self._calc_k_matrix()
 
         # Define matrixes for C to A transformation:
-        tja = np.eye(el_len+2)[:-1, ]
-        tjp1a = np.eye(el_len+2, k=1)[:-1, ]
-
+        tja = np.eye(el_len + 2)[:-1,]
+        tjp1a = np.eye(el_len + 2, k=1)[:-1,]
 
         # Define spline coefficients
         e_mat0 = tja
         e_mat1 = np.dot(tja, k_matrix)
-        e_mat2 = 3 * np.dot(c_mat3**2, (tjp1a-tja)) - \
-                            np.dot(np.dot(c_mat3, (tjp1a + 2 * tja)), k_matrix)
-        e_mat3 = 2 * np.dot(c_mat3**3, (tja-tjp1a)) + \
-                            np.dot(np.dot(c_mat3**2, (tjp1a + tja)), k_matrix)
+        e_mat2 = 3 * np.dot(c_mat3**2, (tjp1a - tja)) - np.dot(
+            np.dot(c_mat3, (tjp1a + 2 * tja)), k_matrix
+        )
+        e_mat3 = 2 * np.dot(c_mat3**3, (tja - tjp1a)) + np.dot(
+            np.dot(c_mat3**2, (tjp1a + tja)), k_matrix
+        )
 
         return e_mat0, e_mat1, e_mat2, e_mat3
 
 
-def estimate_csd(lfp, coord_electrode, sigma, method='standard', diam=None,
-                 h=None, sigma_top=None, tol=1E-6, num_steps=200,
-                 f_type='identity', f_order=None):
+def estimate_csd(
+    lfp,
+    coord_electrode,
+    sigma,
+    method="standard",
+    diam=None,
+    h=None,
+    sigma_top=None,
+    tol=1e-6,
+    num_steps=200,
+    f_type="identity",
+    f_order=None,
+):
     """
     Estimates current source density (CSD) from local field potential (LFP)
     recordings from multiple depths of the cortex.
@@ -1134,23 +1266,29 @@ def estimate_csd(lfp, coord_electrode, sigma, method='standard', diam=None,
     plt.show()
     """
 
-    supported_methods = ('standard', 'delta', 'step', 'spline')
-    icsd_methods = ('delta', 'step', 'spline')
+    supported_methods = ("standard", "delta", "step", "spline")
+    icsd_methods = ("delta", "step", "spline")
 
     if method not in supported_methods:
-        print("Pamareter `method` must be either of {}".format(
-            ", ".join(supported_methods)))
+        print(
+            "Pamareter `method` must be either of {}".format(
+                ", ".join(supported_methods)
+            )
+        )
         raise ValueError
     elif method in icsd_methods and diam is None:
-        print("Parameter `diam` must be specified for iCSD methods: {}".format(
-              ", ".join(icsd_methods)))
+        print(
+            "Parameter `diam` must be specified for iCSD methods: {}".format(
+                ", ".join(icsd_methods)
+            )
+        )
         raise ValueError
 
     if not isinstance(lfp, neo.AnalogSignal):
-        print('Parameter `lfp` must be neo.AnalogSignal')
+        print("Parameter `lfp` must be neo.AnalogSignal")
         raise TypeError
 
-    if f_type != 'identity' and f_order is None:
+    if f_type != "identity" and f_order is None:
         print("The order of {} filter must be specified".format(f_type))
         raise ValueError
 
@@ -1158,32 +1296,35 @@ def estimate_csd(lfp, coord_electrode, sigma, method='standard', diam=None,
     if sigma_top is None:
         sigma_top = sigma
 
-    arg_dict = {'lfp': lfp_pqarr,
-                'coord_electrode': coord_electrode,
-                'sigma': sigma,
-                'f_type': f_type,
-                'f_order': f_order,
-                }
-    if method == 'standard':
+    arg_dict = {
+        "lfp": lfp_pqarr,
+        "coord_electrode": coord_electrode,
+        "sigma": sigma,
+        "f_type": f_type,
+        "f_order": f_order,
+    }
+    if method == "standard":
         csd_estimator = StandardCSD(**arg_dict)
     else:
-        arg_dict['diam'] = diam
-        arg_dict['sigma_top'] = sigma_top
-        if method == 'delta':
+        arg_dict["diam"] = diam
+        arg_dict["sigma_top"] = sigma_top
+        if method == "delta":
             csd_estimator = DeltaiCSD(**arg_dict)
         else:
-            arg_dict['tol'] = tol
-            if method == 'step':
-                arg_dict['h'] = h
+            arg_dict["tol"] = tol
+            if method == "step":
+                arg_dict["h"] = h
                 csd_estimator = StepiCSD(**arg_dict)
             else:
-                arg_dict['num_steps'] = num_steps
+                arg_dict["num_steps"] = num_steps
                 csd_estimator = SplineiCSD(**arg_dict)
     csd_pqarr = csd_estimator.get_csd()
     csd_pqarr_filtered = csd_estimator.filter_csd(csd_pqarr)
-    csd = neo.AnalogSignal(csd_pqarr.T, t_start=lfp.t_start,
-                                         sampling_rate=lfp.sampling_rate)
-    csd_filtered = neo.AnalogSignal(csd_pqarr_filtered.T, t_start=lfp.t_start,
-                                         sampling_rate=lfp.sampling_rate)
+    csd = neo.AnalogSignal(
+        csd_pqarr.T, t_start=lfp.t_start, sampling_rate=lfp.sampling_rate
+    )
+    csd_filtered = neo.AnalogSignal(
+        csd_pqarr_filtered.T, t_start=lfp.t_start, sampling_rate=lfp.sampling_rate
+    )
 
     return csd, csd_filtered