+++ release_notes.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### changed
+- destripe: `k_filter` argument: feeding None to the kfilter does not apply any spatial filter. It is also possible give a function to apply to the data.
+
 ## [1.7.1] - 2025-05-22
 ### fixed
 - remove the offset introduced to sync_timestamps that causes errors downstream due to in place assignment